use XLA tensors instead of meta tensors to trace the offloading graph
This supports hardcoded XLA device names in the input graph (e.g. due
to casts). We need to trace the graph correspondingly with XLA devices.

Why does a cast involve devices? Looks like the ATen ops doesn't include
cast, only `_to_copy`: https://pytorch.org/docs/stable/torch.compiler_ir.html

+++ test_llm_offload.py
@@ -66,7 +66,7 @@ def main(num_layers: int, profile_name: str, num_steps: int, spmd: bool,
       use_flash_attention=flash_attention)
   config.intermediate_size = 4096
   config.vocab_size = 8192
-  model = DecoderOnlyModel(config=config).to(device)
+  model = DecoderOnlyModel(config=config).bfloat16().to(device)
   batch_size = 32
   sequence_length = 1024
